refactor: EventsRepositoryのRequestクラスとResponseクラスを分離
