+++ lib/pages/events_pages/event_list_view.dart
@@ -1,4 +1,5 @@
 import 'package:event_follow/models/controllers/events_controller/events_controller.dart';
+import 'package:event_follow/models/repositories/events/events_api_request.dart';
 import 'package:event_follow/models/repositories/events/events_repository.dart';
 import 'package:event_follow/pages/events_pages/event_content_view.dart';
 import 'package:event_follow/pages/events_pages/event_empty_view.dart';
